kids_books/builder/requirements.txt
fastapi
uvicorn
python-dotenv
openai 